Move bundled backend into dedicated directory
