+++ .gitignore
@@ -13,6 +13,7 @@ lerna-debug.log
 coverage
 errorShots
 examples/*/src-gen
+examples/*/backend
 examples/*/gen-webpack.config.js
 examples/*/gen-webpack.node.config.js
 examples/*/.theia
